クラスの役割ごとに別ディレクトリファイルに分割
